fix(reliability): float counters so decay works below 10 counts
Integer ROUND made every counter <= 10 a fixed point: records decayed
down to ~10 and froze, a seeded record's cancel_rate inflated ~10x as
the denominator hit its floor, and a single recorded no-show
(scheduled=1, observed=0) scored risk 1.0 permanently — exactly the
'bad data never fades' failure decay was built to prevent.

Counters are Float now; decay multiplies without rounding, rows fading
below 0.5 scheduled departures are purged, and reads below the same
threshold fall back to the neutral prior. PostgreSQL migration SQL
recorded in TODO.md and applied to the local Docker DB.

Found by the eighth audit pass (both reviewers, simulation-confirmed).

+++ db/models.py
@@ -134,10 +134,13 @@ class ReliabilityRecord(Base):
     # schedule, "observed" = built from real GTFS-RT observations only,
     # "mixed" = seeded record that has since absorbed real observations.
     source = Column(String, nullable=False, default="observed")
-    observed_departures = Column(Integer, default=0)
-    scheduled_departures = Column(Integer, default=0)
-    total_delay_seconds = Column(Integer, default=0)
-    cancellation_count = Column(Integer, default=0)
+    # Float, not Integer: the daily exponential decay multiplies these by
+    # ~0.95 — integer rounding made every value <= 10 a fixed point that
+    # never decayed, permanently freezing sparse (often bad) records.
+    observed_departures = Column(Float, default=0)
+    scheduled_departures = Column(Float, default=0)
+    total_delay_seconds = Column(Float, default=0)
+    cancellation_count = Column(Float, default=0)
     window_start_date = Column(String)  # YYYYMMDD
     window_end_date = Column(String)    # YYYYMMDD
     updated_at = Column(String)         # ISO 8601 timestamp
